[rename🚚] ServiceApi 모듈 추가로 인한 Domain 모듈내의 단순 파일이동(#CUL-149)
